+++ packages/adapter-api/index.ts
@@ -26,4 +26,3 @@ export * from './src/change_group'
 export * from './src/utils'
 export * from './src/error'
 export * from './src/comparison'
-export * from './src/reference_comparison'
